+++ src/components/amount-input/index.tsx
@@ -1,0 +1,1 @@
+export { AmountInput } from '@/components/amount-input/AmountInput';
