+++ src/components/conversion-info/index.tsx
@@ -1,0 +1,1 @@
+export { ConversionInfo } from '@/components/conversion-info/ConversionInfo';
