+++ src/components/currency-input/index.tsx
@@ -1,0 +1,1 @@
+export { CurrencyInput } from '@/components/currency-input/CurrencyInput';
