+++ src/components/icons/SwapCurrencyIcon.tsx
