+++ src/components/info-tooltip/index.tsx
@@ -1,0 +1,1 @@
+export { InfoTooltip } from '@/components/info-tooltip/InfoTooltip';
