+++ src/components/result/index.tsx
@@ -1,0 +1,1 @@
+export { Result } from '@/components/result/Result';
