+++ src/components/swap-button/SwapButton.tsx
@@ -1,4 +1,4 @@
-import { SwapCurrencyIcon } from './icons/swap-currency-icon';
+import { SwapCurrencyIcon } from '../icons/swap-currency-icon';
 
 export function SwapButton({ onSwap }: { onSwap: () => void }) {
   return (
