+++ src/components/swap-button/index.tsx
@@ -1,0 +1,1 @@
+export { SwapButton } from '@/components/swap-button/SwapButton';
