+++ src/components/toast/index.tsx
@@ -1,0 +1,1 @@
+export { Toast } from '@/components/toast/Toast';
