+++ src/hooks/use-currencies/index.ts
@@ -1,0 +1,1 @@
+export { useCurrencies } from '@/hooks/use-currencies/useCurrencies';
